Rename frame_buffer.{c,h} to frame.{c,h}
