+++ src/compositor/CMakeLists.txt
@@ -3,7 +3,7 @@
 #
 set(SOURCE_FILES
     buffer/buffer.c
-    buffer/frame_buffer.c
+    buffer/frame.c
     buffer/image_buffer.c
     cursor.c
     framebuffer_device.c
