+++ src/compositor/buffer/frame.c
@@ -38,7 +38,7 @@
 #include "logger/module.h"
 #include "objects/object.h"
 
-#include "compositor/buffer/frame_buffer.h"
+#include "compositor/buffer/frame.h"
 #include "compositor/buffer/image_buffer.h"
 #include "compositor/internal_context.h"
