+++ src/compositor/cursor.c
@@ -33,7 +33,7 @@
 #include "objects/object.h"
 #include "util/arithmetical.h"
 
-#include "compositor/buffer/frame_buffer.h"
+#include "compositor/buffer/frame.h"
 #include "compositor/cursor.h"
 #include "compositor/framebuffer_device.h"
 #include "compositor/internal_context.h"
